# rentalApp/src/main/resources/app-api-swagger.json.yml
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/user/settings": {
      "get": {
        "tags": [
          "Settings Controller"
        ],
        "summary": "User settings form",
        "operationId": "settingsProfile",
        "parameters": [
          {
            "name": "userDetails",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UserDetails"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Settings Controller"
        ],
        "summary": "Update user",
        "operationId": "update",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreateEditDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "UserCreateEditDto": {
        "required": [
          "address",
          "password",
          "username"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "maxLength": 20,
            "minLength": 8,
            "pattern": "^(?=.*\\d)(?=.*[a-z])(?=.*[A-Z]).{8,}$",
            "type": "string"
          },
          "birthDate": {
            "type": "string",
            "format": "date"
          },
          "firstname": {
            "maxLength": 20,
            "minLength": 2,
            "pattern": "[A-Za-z]+",
            "type": "string"
          },
          "lastname": {
            "maxLength": 25,
            "minLength": 2,
            "pattern": "[A-Za-z]+",
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "pesel": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "profilePicture": {
            "type": "string",
            "format": "binary"
          },
          "role": {
            "type": "string",
            "enum": [
              "USER",
              "ADMIN"
            ]
          },
          "emailVerified": {
            "type": "boolean"
          },
          "emailVerificationToken": {
            "type": "string"
          }
        }
      },
...
      "GrantedAuthority": {
        "type": "object",
        "properties": {
          "authority": {
            "type": "string"
          }
        }
      },
      "UserDetails": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "username": {
            "type": "string"
          },
          "authorities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GrantedAuthority"
            }
          },
          "accountNonExpired": {
            "type": "boolean"
          },
          "credentialsNonExpired": {
            "type": "boolean"
          },
          "accountNonLocked": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}